Expose libavutil Error.h macros to Swift
https://github.com/tanersener/mobile-ffmpeg/issues/289

Complex (function) c macros are not exposed to Swift.
In order to access these errors, we need to expose c functions which swift can access.

By prefixing an underscore on the function name, and using `__attribute__((swift_name("foo()")))` we can make the swift names almost identical to the c macro names without conflicting with the original macros.